requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
ffmpeg-python==0.2.0
future==0.18.2
idna==3.2
python-telegram-bot==13.7
pytube==11.0.1
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
